Handle CHECKBOX_CHANGED event in MainMenu event handler (fixes #336)
Previously, changing a setting in the Settings tab of the main menu
would be ignored when one immediately leaves the program. Because the
event handler OnEvent() didn't check for CHECKBOX_CHANGED events, the
settings were only read after some other GUI event happened, like
changing to another tab, for example. As a result, the change would
not be written to the config file, because the internal settings
didn't change.